+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+*Version 1.15.1* **November 25, 2024**
+- Removed the Date extension for Strideable, as it has been in the various OSes, for a while.
+- Upped the base OS version for iOS, TVOS, and WatchOS, slightly.
+
 *Version 1.15.0* **November 20, 2024**
 - Added the `RVS_WeakObjectReference` struct.
 - Tweaked the tests to work properly on all platforms.